준비작업: js 파일 분리
