+++ index.html
@@ -7,6 +7,7 @@
     <link href="style/css/main.css" rel="stylesheet" />
     <link href="style/css/color.css" rel="stylesheet" />
     <link rel="icon" href="style/image/icon.png" />
+    <script src="./src/main.js" type="module"></script>
     <body>
       <nav class="menubar-top">
         <img id="representive-mark" src="style/image/icon.png" />
@@ -22,8 +23,6 @@
       </nav>
 
       <div class="content" id="eiga"></div>
-
-      <script src="main.js"></script>
     </body>
   </head>
 </html>
