+++ style/css/color.css
@@ -3,12 +3,3 @@
   --light-color: #cedff6;
   --white-color: white;
 }
-@font-face {
-  font-family: "Pretendard";
-  src: url("../fonts/Pretendard-Regular.woff2") format("woff2");
-}
-
-@font-face {
-  font-family: "Pretendard-extrabold";
-  src: url("../fonts/Pretendard-ExtraBold.woff2") format("woff2");
-}
